# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Insurance Broker System (IBS) API
  description: |
    RESTful API for the Insurance Broker System - a multi-tenant insurance agency management platform.

    ## Authentication
    All endpoints (except authentication) require a valid JWT Bearer token obtained via the `/auth/login` endpoint.

    ## Multi-Tenancy
    The system is multi-tenant. Tenant context is determined from the JWT token claims.

    ## Error Responses
    All error responses follow a consistent format with an `error` field containing the error message.
  version: 1.0.0
  contact:
    name: IBS API Support
  license:
    name: Proprietary
servers:
  - url: https://localhost:7001/api/v1
    description: Development server
  - url: https://api.ibs.example.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /documents/upload:
    post:
      tags:
        - Documents
      summary: Upload a document
      description: Uploads a file with associated metadata. Returns the new document ID.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - entityType
                - category
              properties:
                file:
                  type: string
                  format: binary
                entityType:
                  $ref: '#/components/schemas/DocumentEntityType'
                category:
                  $ref: '#/components/schemas/DocumentCategory'
                entityId:
                  type: string
                  format: uuid
                  description: Optional ID of the linked entity (policy, client, claim, carrier)
                description:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      operationId: getDocumentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailDto'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Documents
      summary: Archive a document
      description: Soft-deletes a document by marking it as archived.
      operationId: archiveDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document archived successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download document
      description: Returns a redirect (302) to a temporary pre-signed SAS URL for the document blob.
      operationId: downloadDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Redirect to pre-signed download URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # Document Template Endpoints
  # ============================================
  /documents/templates:
    get:
      tags:
        - Document Templates
      summary: List document templates
      operationId: listDocumentTemplates
      parameters:
        - name: searchTerm
          in: query
          schema:
            type: string
        - name: templateType
          in: query
          schema:
            $ref: '#/components/schemas/TemplateType'
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated template list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTemplateSearchResultDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Document Templates
      summary: Create document template
      operationId: createDocumentTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/templates/{id}:
    get:
      tags:
        - Document Templates
      summary: Get template by ID
      operationId: getDocumentTemplateById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Template detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTemplateDetailDto'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Document Templates
      summary: Update document template
      operationId: updateDocumentTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentTemplateRequest'
      responses:
        '204':
          description: Template updated
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/templates/{id}/activate:
    post:
      tags:
        - Document Templates
      summary: Activate a template
      operationId: activateDocumentTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Template activated
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/templates/{id}/deactivate:
    post:
      tags:
        - Document Templates
      summary: Deactivate a template
      operationId: deactivateDocumentTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Template deactivated
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # COI Generation Endpoints
  # ============================================
  /documents/generate-coi:
    post:
      tags:
        - COI Generation
      summary: Generate Certificate of Insurance
      description: |
        Renders the specified Handlebars template with live policy data from the database
        (client name, carrier name, coverages, broker name, etc.) using QuestPDF to produce
        a PDF, stores it in Azure Blob Storage, and returns the new document ID.
      operationId: generateCOI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCOIRequest'
      responses:
        '201':
          description: COI generated and stored as a document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Template or policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message

    # ============================================
    # Authentication Schemas
    # ============================================
...
    DocumentCategory:
      type: string
      enum:
        - Policy
        - Endorsement
        - COI
        - ClaimReport
        - KYC
        - Invoice
        - Other
    DocumentEntityType:
      type: string
      enum:
        - Policy
        - Client
        - Claim
        - Carrier
        - General
    TemplateType:
      type: string
      enum:
        - CertificateOfInsurance
        - PolicySummary
...
    DocumentDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        entityType:
          $ref: '#/components/schemas/DocumentEntityType'
        entityId:
          type: string
          format: uuid
          nullable: true
        fileName:
          type: string
        contentType:
          type: string
        fileSizeBytes:
          type: integer
          format: int64
        blobKey:
          type: string
        category:
          $ref: '#/components/schemas/DocumentCategory'
        version:
          type: integer
        isArchived:
          type: boolean
        uploadedBy:
          type: string
        description:
          type: string
          nullable: true
        uploadedAt:
          type: string
          format: date-time
...
    DocumentTemplateListItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        templateType:
          $ref: '#/components/schemas/TemplateType'
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
    DocumentTemplateDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        templateType:
          $ref: '#/components/schemas/TemplateType'
        content:
          type: string
          description: Handlebars HTML template content
        isActive:
          type: boolean
        version:
          type: integer
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentTemplateSearchResultDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentTemplateListItemDto'
        totalCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    CreateDocumentTemplateRequest:
      type: object
      required:
        - name
        - templateType
        - content
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        templateType:
          $ref: '#/components/schemas/TemplateType'
        content:
          type: string
          description: Handlebars HTML template content
    UpdateDocumentTemplateRequest:
      type: object
      required:
        - name
        - content
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        content:
          type: string
          description: Handlebars HTML template content
    GenerateCOIRequest:
      type: object
      required:
        - templateId
        - policyId
      properties:
        templateId:
          type: string
          format: uuid
          description: ID of an active CertificateOfInsurance template
        policyId:
          type: string
          format: uuid
          description: ID of the policy to pull data from